Notes Delete notes Show past searches

+++ application/views/notes/history.php
@@ -9,33 +9,89 @@
             <span id="createNote"><img src="/images/third_party/icons/Add_icon.png" />
             </span>
         </h2>
+        <?php if (!empty($search)) {
+            ?>
+            <h3 id="SearchCriteria">
+                Search Text: <?php echo $search[0]["text"]; ?> 
+                From Date: <?php echo strtotime($search[0]["start_date"]) . " | " . (strtotime($search[0]["start_date"]) == 0) ? "None" : $search[0]["start_date"]; ?> 
+                To Date: <?php echo strtotime($search[0]["end_date"]) . " | " . (strtotime($search[0]["end_date"]) == 0) ? "None" : $search[0]["end_date"]; ?> 
+
+            </h3>
+            <?php
+        }
+        ?>
         <div id="searchNotesHistory" >
             <div id="searchForm" class="notesHistoryHidden notesHistory">
                 <form action="/notes/history/search" method="post" accept-charset="utf-8" id="SearchNoteForm">
                     <div class="row">
                         <div class="large-12 columns" >
-                            <label for="title">Text
-                                <input type="text" name="searchText" placeholder="Topic, subject, content or tag text you wish to search for" value="" />
-                            </label>
-                        </div>
-                    </div>
-                    <div class="row">
-                        <div class="large-2 columns" >
-                            <label>Date From  
-                                <input id="fromDate" type="text" placeholder="<?php echo date('Y/m/d H:i:s', mktime(0, 0, 0, date("m") - 1, date("d"), date("Y"))); ?>" name="fromDate" />
-                            </label>
-                        </div>
-                        <div class="large-2 columns">
-                            <label>To 
-                                <input id="toDate" type="text" placeholder="<?php echo date('Y/m/d H:i:s'); ?>" name="toDate" />
-                            </label>
-                        </div>
-                        <div class="large-8 columns" ></div>
-                    </div>
-                    <div class="row">
-                        <div class="large-12 columns" >
-                            <br/>
-                            <input type="submit" value="Search" text="Search" class="button" />
+                            <div class="row">
+                                <div class="large-6 columns" >
+                                    <label for="title">Text
+                                        <input type="text" name="searchText" placeholder="Topic, subject, content or tag text you wish to search for" value="" />
+                                    </label>
+                                    <div class="row">
+                                        <div class="large-2 columns" >
+                                            <label>Date From  
+                                                <input id="fromDate" type="text" placeholder="<?php echo date('Y/m/d H:i:s', mktime(0, 0, 0, date("m") - 1, date("d"), date("Y"))); ?>" name="fromDate" />
+                                            </label>
+                                        </div>
+                                        <div class="large-2 columns">
+                                            <label>To 
+                                                <input id="toDate" type="text" placeholder="<?php echo date('Y/m/d H:i:s'); ?>" name="toDate" />
+                                            </label>
+                                        </div>
+                                        <div class="large-8 columns" ></div>
+                                    </div>
+                                    <div class="row">
+                                        <div class="large-12 columns" >
+                                            <br/>
+                                            <input type="submit" value="Search" text="Search" class="button" />
+                                        </div>
+                                    </div>
+                                </div>
+                                <div class="large-6 columns" >
+                                    <?php
+                                    if (empty($searches)) {
+                                        echo "<h3>No Prior searches.</h3>";
+                                    } else {
+                                        ?>
+                                        <div class="row">
+                                            <div class="large-12 columns" >
+                                                <h3>Past searches</h3>
+                                            </div>
+                                        </div>
+                                        <div class="row">
+                                           <div class='large-4 columns' >
+                                               Text
+                                            </div>
+                                           <div class='large-4 columns' >
+                                               From Date
+                                            </div>
+                                           <div class='large-4 columns' >
+                                               To Date
+                                            </div>
+                                        </div>
+                                        <?php    
+                                        foreach ($searches as $kk => $vv) {
+                                            ?>
+                                            <div class="row">
+                                                <div class='large-4 columns' >
+                                                    <a href='/notes/history/search/<?php echo $vv["id"]; ?>' ><?php echo $vv["text"]; ?></a>
+                                                </div>
+                                                <div class='large-4 columns' >
+                                                    <a href='/notes/history/search/<?php echo $vv["id"]; ?>' ><?php echo (($vv["start_date"] == "0000-00-00 00:00:00") ? "None" : $vv["start_date"]); ?></a>
+                                                </div>
+                                                <div class='large-4 columns' >
+                                                    <a href='/notes/history/search/<?php echo $vv["id"]; ?>' ><?php echo (($vv["end_date"] == "0000-00-00 00:00:00") ? "None" : $vv["end_date"]) ; ?></a>
+                                                </div>
+                                            </div>
+                                            <?php
+                                        }
+                                    }
+                                    ?>
+                                </div>
+                            </div>
                         </div>
                     </div>
                 </form>
@@ -49,7 +105,7 @@
         <div id="Notes">
             <h3>Notes</h3>
             <div class="pagination-centered">
-            <?php echo $this->pagination->create_links(); ?>
+                <?php echo $this->pagination->create_links(); ?>
             </div>
 
             <?php
@@ -100,7 +156,7 @@
                         <div id="action_content" class="note_action">
                             <a href="/notes/edit/<?php echo $v["id"]; ?>">edit</a>
                             &nbsp;&nbsp;|&nbsp;&nbsp;
-                            <a href="/notes/delete/<?php echo $v["id"]; ?>" >delete</a>                            
+                            <a href="/notes/delete/<?php echo $v["id"]; ?>"  onclick="return confirm_delete()">delete</a>                            
                         </div>
                     </div>
                     <div style="clear:both;" />
